Extracted GameData.
