+++ TombRaider1996/TR1.csproj
@@ -53,6 +53,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="GameData.cs" />
     <Compile Include="GameMemory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ComponentSettings.cs">
